Fixed error

+++ data/mri_data.py
@@ -215,7 +215,7 @@ class SelectiveSliceData_Val(torch.utils.data.Dataset):
         print('total_files:')
         print(len(f))
 
-        for fname in f[1:len(f)*0.1]:
+        for fname in f[1:int(len(f)*0.1)]:
             kspace = h5py.File(fname, 'r')['kspace']
             with h5py.File(fname, 'r') as data:
                 if (data.attrs['acquisition'] == 'AXT2'):
